Refactoring therapies in seperated file
